requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.4.4
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
cssselect2==0.9.0
Django==6.0.2
django-admin-sortable2==2.3.1
django-environ==0.13.0
django-jazzmin==3.0.2
django-tailwind==4.4.2
fonttools==4.62.0
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
pyphen==0.17.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
zopfli==0.4.1
